requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-image-optimizer==1.0.1
djangorestframework==3.11.0
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
future==0.18.2
idna==2.10
kombu==4.6.11
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-resize-image==1.1.19
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tinify==1.5.1
urllib3==1.25.9
vine==1.3.0
